Refactored graph helper functions from Svg.Database to Models.Graph (#1716)
