+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 - Renamed usages of the word "room" to "location" in the codebase to better reflect the data represented
 - Added test cases for JSON parsing of Meeting data type in `backend-test/Database/TablesTests.hs`
 - Added test cases for JSON parsing of Time' data type in `backend-test/Database/TablesTests.hs`
+- Refactor graph helper functions from `app/Svg/Database.hs` to `app/Models/Graph.hs`
+- Refactor functions for performing matrix operations from `app/Svg/Parser.hs` to `app/Util/Matrix.hs`
 
 ## [0.7.2] - 2025-12-10
