+++ app/Svg/Parser.hs
@@ -5,7 +5,7 @@ This module is reponsible for parsing the SVG files exported from Inkscape.
 It also currently acts as a main driver for the whole graph pipeline:
 
 1. Parsing the raw SVG files
-2. Inserting them into the database (see "Svg.Database")
+2. Inserting them into the database (see "Models.Graph")
 3. Retrieving the database values and generating a new SVG file
    (See "Svg.Builder" and "Svg.Generator")
 
@@ -14,19 +14,19 @@ directly to the client when viewing the @/graph@ page.
 -}
 
 module Svg.Parser
-    (parsePrebuiltSvgs, parseDynamicSvg, matrixPointMultiply) where
+    (parsePrebuiltSvgs, parseDynamicSvg) where
 
 import Config (graphPath, runDb)
 import Control.Monad.IO.Class (liftIO)
 import Data.Char (isSpace)
-import Data.List as List
+import qualified Data.List as List
 import Data.Maybe (fromMaybe)
 import qualified Data.Text as T
 import Data.Text.IO as T (readFile)
 import Database.DataType
 import Database.Persist.Sqlite
 import Database.Tables hiding (graphHeight, graphWidth, paths, shapes, texts)
-import Svg.Database (deleteGraph, insertElements, insertGraph)
+import Models.Graph (deleteExistingGraph, insertElements)
 import qualified Text.HTML.TagSoup as TS hiding (fromAttrib)
 import Text.HTML.TagSoup (Tag)
 import qualified Text.Parsec as P
@@ -34,6 +34,7 @@ import Text.Parsec ((<|>))
 import Text.Parsec.String (Parser)
 import Text.Read (readMaybe)
 import Util.Helpers
+import Util.Matrix (matrixMultiply)
 
 
 parsePrebuiltSvgs :: IO ()
@@ -76,22 +77,12 @@ performParse graphName inputFilename = do
     graphFile <- liftIO $ T.readFile (gPath ++ inputFilename)
     performParseFromMemory graphName graphFile False
 
--- | Deletes the graph with the given name from the database if it exists.
-deleteExistingGraph :: T.Text -> SqlPersistM ()
-deleteExistingGraph graphName = do
-  graphEnt :: (Maybe (Entity Graph)) <- selectFirst [GraphTitle ==. graphName] []
-  case graphEnt of
-    Just graph -> do
-      let gId = entityKey graph
-      deleteGraph gId
-    Nothing -> pure ()
-
 performParseFromMemory :: T.Text -- ^ The title of the graph
                        -> T.Text -- ^ Filename of the SVG to parse
                        -> Bool -- ^ True if SVG is dynamically generated
                        -> SqlPersistM ()
 performParseFromMemory graphName graphSvg isDynamic = do
-    key <- insertGraph graphName graphWidth graphHeight isDynamic
+    key <- insert (Graph graphName graphWidth graphHeight isDynamic)
     let parsedGraph = parseSvg key graphSvg
     insertElements parsedGraph
         where (graphWidth, graphHeight) = parseSizeFromSvg graphSvg
@@ -667,23 +658,6 @@ updateShape fill r =
 -- | Adds two tuples together.
 addTuples :: Point -> Point -> Point
 addTuples (a,b) (c,d) = (a + c, b + d)
-
--- | Apply a matrix transformation to a point.
--- The matrix must have dimensions 3x3, representing a transformation for a two-dimensional point,
--- i.e., the transformation in the z-component is ignored, so the third row is expected to be [0,0,1]
-matrixPointMultiply :: Matrix -> Point -> Point
-matrixPointMultiply matrix (x, y) =
-    case matrix of
-        [[a, b, tx], [c, d, ty], _] -> (a * x + b * y + tx, c * x + d * y + ty)
-        _ -> error "Matrix must be 3x3 for point transformation."
-
--- | Multiplies two 3x3 matrices together.
-matrixMultiply :: Matrix -> Matrix -> Matrix
-matrixMultiply m1 m2 = [[dotProduct row col | col <- transpose m2] | row <- m1]
-
--- | Computes the dot product of two vectors.
-dotProduct :: Vector -> Vector -> Double
-dotProduct v1 v2 = sum $ zipWith (*) v1 v2
 
 -- | Helper to remove leading and trailing whitespace.
 trim :: T.Text -> T.Text
