+++ courseography.cabal
@@ -49,12 +49,12 @@ library
     Models.Program,
     Scripts,
     Svg.Builder,
-    Svg.Database,
     Svg.Generator,
     Svg.Parser,
     Util.Blaze,
     Util.Happstack,
     Util.Helpers,
+    Util.Matrix,
     WebParsing.ArtSciParser,
     WebParsing.ParsecCombinators,
     WebParsing.PostParser,
@@ -201,13 +201,13 @@ executable courseography
     Scripts,
     Server,
     Svg.Builder,
-    Svg.Database,
     Svg.Generator,
     Svg.Parser,
     Util.Blaze,
     Util.Documentation,
     Util.Happstack,
     Util.Helpers,
+    Util.Matrix,
     WebParsing.ArtSciParser,
     WebParsing.Ligature,
     WebParsing.ParsecCombinators,
